Automatically decode sanitized completenames
fixes #21247
